requirements.txt
Django==3.2.1
django-filter==2.4.0
django-test-without-migrations==0.6
djangorestframework==3.12.4
Markdown==3.3.4
pytest-django==4.2.0
pytest-mock==3.6.0
python-decouple==3.4
model-bakery==1.3.1
